refactor: update message box use hooks
